Revert "fix: cherry-pick the narrowing conversion fix code of the net, storage, pstd modules(no problem local test)"
This reverts commit ec537fea9d58f1b404b374ecb8859db1f7a1f7c5.

+++ src/net/src/bg_thread.cc
@@ -26,8 +26,8 @@ void BGThread::Schedule(void (*function)(void*), void* arg) {
 
 void BGThread::QueueSize(int* pri_size, int* qu_size) {
   std::lock_guard lock(mu_);
-  *pri_size = static_cast<int32_t>(timer_queue_.size());
-  *qu_size = static_cast<int32_t>(queue_.size());
+  *pri_size = timer_queue_.size();
+  *qu_size = queue_.size();
 }
 
 void BGThread::QueueClear() {
